Ensure websocket termination handled correctly
Currently when a websocket is terminated on the client side, the logic
in EventNotifier continues to run. This is because the WebSocketError
exceptions are detected in the properties/interface handlers where
that exception is just ignored.

This exception can not be sent back up to the main loop because
it goes through the dbus libraries which will fail when it
sees it.

Use a class variable to indicate when a websocket error has
been hit and break out of the loop. Also, remove signal handlers.

Change-Id: I068f19fd75aa03fb71d98e9ff75f596941c8622f
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Signed-off-by: Deepak Kodihalli <dkodihal@in.ibm.com>